+++ client/src/components/app.js
@@ -4,7 +4,7 @@ import { Route } from "react-router-dom";
 import Header from "./Header";
 import Footer from "./Footer";
 
-import Leaderboard from "../containers/leaderboard";
+import Dashboard from "../containers/leaderboard";
 import Home from "../containers/Home";
 import AddQuestion from "../containers/addQuestion";
 import Question from "../containers/question"
@@ -18,7 +18,7 @@ export default class App extends Component {
                 <Header />
                 <div className="container">
                     <Route exact path="/" component={Home} />
-                    <Route exact path="/leaderboard" component={Leaderboard} />
+                    <Route exact path="/dashboard" component={Dashboard} />
                     <Route exact path="/addquestion" component={AddQuestion} />
                     <Route exact path="/questions/:id" component={Question}/>
                 </div>