+++ client/src/components/header/index.js
@@ -9,7 +9,7 @@ export default class Header extends Component {
 	constructor(props) {
 		super(props)
 		this.state = {
-			navItems: ["home", "leaderboard", "addquestion"],
+			navItems: ["home", "dashboard", "addquestion"],
 			open: false
 		}
 	}