+++ client/src/containers/home.js
@@ -1,13 +1,14 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
+import Sidebar from "../components/Sidebar";
 
 export default class Home extends Component {
-  render() {
-    return (
-      <div>
-        <h1 className="title">A social media for Software Engineers</h1>
-        <a href="/auth/google">Sign in with Google</a>
-      </div>
-    );
-  }
+	render() {
+		return (
+			<div>
+				<h1 className="title">A social media for Software Engineers</h1>
+				<Sidebar />
+			</div>
+		);
+	}
 }